+++ src/TestUtils/Exceptions/BadMethodException.php
@@ -6,12 +6,12 @@ namespace Raptor\TestUtils\Exceptions;
 use RuntimeException;
 
 /**
- * Exception, that is thrown when _ExtraUtils::invokeMethod_ is called with incorrect method name.
+ * Exception that is thrown when _ExtraUtils::invokeMethod_ is called with incorrect method name.
  *
  * @author Mikhail Kamorin aka raptor_MVK
  *
  * @copyright 2019, raptor_MVK
  */
-class BadMethodException extends RuntimeException
+final class BadMethodException extends RuntimeException
 {
 }