+++ src/TestUtils/Exceptions/DataDirectoryNotFoundException.php
@@ -6,12 +6,12 @@ namespace Raptor\TestUtils\Exceptions;
 use RuntimeException;
 
 /**
- * Exception that is thrown, when directory with data files was not found.
+ * Exception that is thrown when directory with data files was not found.
  *
  * @author Mikhail Kamorin aka raptor_MVK
  *
  * @copyright 2019, raptor_MVK
  */
-class DataDirectoryNotFoundException extends RuntimeException
+final class DataDirectoryNotFoundException extends RuntimeException
 {
 }