+++ src/TestUtils/Exceptions/DataFileNotFoundException.php
@@ -6,12 +6,12 @@ namespace Raptor\TestUtils\Exceptions;
 use RuntimeException;
 
 /**
- * Exception that is thrown, when data file was not found.
+ * Exception that is thrown when data file was not found.
  *
  * @author Mikhail Kamorin aka raptor_MVK
  *
  * @copyright 2019, raptor_MVK
  */
-class DataFileNotFoundException extends RuntimeException
+final class DataFileNotFoundException extends RuntimeException
 {
 }