+++ tests/TestUtils/DataLoader/DirectoryDataLoaderFactoryTests.php
@@ -6,14 +6,14 @@ namespace RaptorTests\TestUtils\DataLoader;
 use PHPUnit\Framework\TestCase;
 use Raptor\TestUtils\DataLoader\DataLoaderFactory;
 use Raptor\TestUtils\DataLoader\DirectoryDataLoaderFactory;
-use Raptor\TestUtils\DataProcessor\TestContainerGeneratorDataProcessor;
+use Raptor\TestUtils\DataProcessor\GeneratorDataProcessor;
 
 /**
  * @author Mikhail Kamorin aka raptor_MVK
  *
  * @copyright 2019, raptor_MVK
  */
-class DirectoryDataLoaderFactoryTests extends TestCase
+final class DirectoryDataLoaderFactoryTests extends TestCase
 {
     /**
      * Checks that factory method _createTestContainerGeneratorDataLoader_ returns DirectoryDataLoader with
@@ -27,6 +27,6 @@ class DirectoryDataLoaderFactoryTests extends TestCase
 
         $actualDataProcessorClass = $directoryDataLoader->getDataProcessorClass();
 
-        static::assertSame(TestContainerGeneratorDataProcessor::class, $actualDataProcessorClass);
+        static::assertSame(GeneratorDataProcessor::class, $actualDataProcessorClass);
     }
 }