+++ tests/TestUtils/DataLoader/ProcessedDataLoaderTests.php
@@ -8,19 +8,22 @@ use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 use Mockery\MockInterface;
 use PHPUnit\Framework\TestCase;
 use Raptor\TestUtils\DataLoader\DataLoader;
+use Raptor\TestUtils\DataLoader\ProcessedDataLoader;
 use Raptor\TestUtils\DataProcessor\DataProcessor;
-use Raptor\TestUtils\DataProcessor\TestContainerGeneratorDataProcessor;
-use Raptor\TestUtils\DataProcessor\TestContainerWrapperDataProcessor;
+use Raptor\TestUtils\DataProcessor\GeneratorDataProcessor;
+use Raptor\TestUtils\DataProcessor\TypeFactory\GetTypeTypeFactory;
+use Raptor\TestUtils\DataProcessor\WrapperDataProcessor;
 use Raptor\TestUtils\Exceptions\DataFileNotFoundException;
 use Raptor\TestUtils\ExtraAssertions;
 use Raptor\TestUtils\WithVFS;
+use function get_class;
 
 /**
  * @author Mikhail Kamorin aka raptor_MVK
  *
  * @copyright 2019, raptor_MVK
  */
-class DataLoaderTests extends TestCase
+final class ProcessedDataLoaderTests extends TestCase
 {
     use MockeryPHPUnitIntegration, ExtraAssertions, WithVFS;
 
@@ -34,19 +37,16 @@ class DataLoaderTests extends TestCase
      * Checks that method _getDataProcessorClass_ returns correct class.
      *
      * @param DataProcessor $dataProcessor
-     * @param string $expectedClass
      *
      * @dataProvider dataProcessorClassDataProvider
      */
-    public function testGetDataProcessorClassReturnsCorrectClass(
-        DataProcessor $dataProcessor,
-        string $expectedClass
-    ): void {
-        $dataLoader = new DataLoader($dataProcessor);
+    public function testGetDataProcessorClassReturnsCorrectClass(DataProcessor $dataProcessor): void
+    {
+        $dataLoader = new ProcessedDataLoader($dataProcessor);
 
         $actualClass = $dataLoader->getDataProcessorClass();
 
-        static::assertSame($expectedClass, $actualClass);
+        static::assertSame(get_class($dataProcessor), $actualClass);
     }
 
     /**
@@ -57,8 +57,8 @@ class DataLoaderTests extends TestCase
     public function dataProcessorClassDataProvider(): array
     {
         return [
-            'wrapper' => [new TestContainerWrapperDataProcessor(), TestContainerWrapperDataProcessor::class],
-            'generator' => [new TestContainerGeneratorDataProcessor(), TestContainerGeneratorDataProcessor::class]
+            'wrapper' => [new WrapperDataProcessor()],
+            'generator' => [new GeneratorDataProcessor(new GetTypeTypeFactory())]
         ];
     }
 
@@ -91,7 +91,7 @@ class DataLoaderTests extends TestCase
         if ($dataProcessorMockCallback !== null) {
             $dataProcessorMockCallback($dataProcessorMock);
         }
-        return new DataLoader($dataProcessorMock);
+        return new ProcessedDataLoader($dataProcessorMock);
     }
 
     /**